refactor(www): add modules/marketing, extract home page
Introduce modules/marketing as the home-page feature and move the
showcase orchestrator (cards.tsx) into it, clearing the last
components -> modules back-edge (cards.tsx imported @/modules/core and
@/modules/create). The reusable CardsGrid + card leaves stay in
components/showcase; only the stateful orchestrator moves.

Extract the home page body from routes/_app/index.tsx into
modules/marketing/home-page.tsx so the route is a thin delegator, and
drop the dead commented-out "Fueled by your stars" CTA block.

cards.tsx is the "Open in v0" bundle entry, so repoint the bundle entry,
the v0 page template, and the materialize harness at the new path, and
add a modules/** @source glob so Tailwind still scans the relocated
(class-heavy) cards.tsx. Regenerate showcase-bundle.ts.
