+++ www/scripts/materialize-bundle.ts
@@ -40,6 +40,7 @@ const GLOBALS_CSS = `@import "tailwindcss";
 
 @source "../registry/**/*.{ts,tsx}";
 @source "../components/**/*.{ts,tsx}";
+@source "../modules/**/*.{ts,tsx}";
 @source "./**/*.{ts,tsx}";
 
 @theme {
@@ -50,7 +51,7 @@ const GLOBALS_CSS = `@import "tailwindcss";
 
 const PAGE_TSX = `"use client";
 
-import Cards from "../components/showcase/cards";
+import Cards from "../modules/marketing/cards";
 import { DesignSystemProvider } from "../modules/core/styles";
 
 export default function Page() {
