+++ www/src/routes/_app/index.tsx
@@ -1,73 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router'
 
-import { LinkButton } from '@/registry/ui/button'
-import { Tooltip, TooltipContent } from '@/registry/ui/tooltip'
-import { Announcement } from '@/components/announcement'
-import { BaseUiIcon } from '@/components/icons/base-ui'
-import { BoltIcon } from '@/components/icons/bolt'
-import { LovableIcon } from '@/components/icons/lovable'
-import { ReactAriaIcon } from '@/components/icons/react-aria'
-import { ReactJsIcon } from '@/components/icons/react-js'
-import { ShadcnIcon } from '@/components/icons/shadcn'
-import { TailwindWordmark } from '@/components/icons/tailwind-wordmark'
-import { TypeScriptIcon } from '@/components/icons/typescript'
-import { V0Icon } from '@/components/icons/v0'
-import { Footer } from '@/components/layout/footer'
-import {
-  AnimatedHeadline,
-  type RotatingTextItem,
-} from '@/components/rotating-text'
-import Cards from '@/components/showcase/cards'
-
-// dotUI's brand font (the "dotUI" wordmark uses Josefin Sans), loaded locally. Weight,
-// letter-spacing and size come from the original headline classes on the <h1>.
-const HEADLINE_STYLE = { fontFamily: 'var(--font-josefin)' }
-
-// Shared connector for the tool / codebase frames. The trailing space is a non-breaking
-// space on purpose: the rotator sizes the connector slot from this text, and a normal
-// trailing space gets trimmed in the flex slot (so "to" would butt against the logo).
-const TO = 'to '
-
-// Swappable destinations in the hero headline — "Ship it ___". The shared "to" lives in
-// `connector`, not the lead, so it renders once in the lead's font and — because the
-// rotator keys the connector slot by this string — stays put across the v0 / bolt.new /
-// Lovable / your-code frames instead of re-animating; only the destination swaps.
-// "anywhere" reads without a connector. v0 / bolt.new / Lovable are their official brand
-// wordmarks (logo-only); the rest are text. Everything inherits the headline color
-// (white); the Lovable heart keeps its brand gradient.
-const EXPORT_TARGETS: RotatingTextItem[] = [
-  {
-    id: 'anywhere',
-    text: 'anywhere',
-    segments: [{ text: 'anywhere', className: 'font-bold italic' }],
-  },
-  {
-    id: 'v0',
-    text: 'to v0',
-    connector: TO,
-    segments: [{ icon: <V0Icon className="h-[0.58em] w-auto" /> }],
-  },
-  {
-    id: 'bolt',
-    text: 'to bolt.new',
-    connector: TO,
-    segments: [
-      { icon: <BoltIcon className="h-[0.72em] w-auto translate-y-[0.09em]" /> },
-    ],
-  },
-  {
-    id: 'lovable',
-    text: 'to Lovable',
-    connector: TO,
-    segments: [{ icon: <LovableIcon className="h-[0.72em] w-auto" /> }],
-  },
-  {
-    id: 'codebase',
-    text: 'to your code',
-    connector: TO,
-    segments: [{ text: 'your code', className: 'font-bold italic' }],
-  },
-]
+import { HomePage } from '@/modules/marketing/home-page'
 
 export const Route = createFileRoute('/_app/')({
   head: () => ({
@@ -76,163 +9,3 @@ export const Route = createFileRoute('/_app/')({
   }),
   component: HomePage,
 })
-
-function HomePage() {
-  return (
-    <div className="min-h-[calc(100vh-var(--header-height))]">
-      <main>
-        {/* Hero section */}
-        <section className="container flex flex-col pt-6 sm:pt-8 md:pt-12">
-          <div className="flex flex-col items-center gap-3 text-center md:gap-4">
-            <Announcement />
-            <h1
-              aria-label="Build your design system. Ship it to v0, bolt.new, Lovable, your code, or anywhere."
-              className="text-3xl leading-tight tracking-tighter text-balance max-lg:font-medium md:text-4xl lg:text-5xl"
-            >
-              <span aria-hidden="true">
-                <AnimatedHeadline
-                  lead="Build your design system. Ship it"
-                  items={EXPORT_TARGETS}
-                  trailing="."
-                  wordStyle={HEADLINE_STYLE}
-                />
-              </span>
-            </h1>
-            <p className="max-w-2xl text-base text-balance text-fg-muted sm:text-lg">
-              Beautiful components, accessibility out of the box, composition,
-              and more.
-              <br />
-              Powered by{' '}
-              <ReactAriaIcon className="inline-flex h-[0.85em] w-auto translate-y-[-0.08em]" />{' '}
-              <span className="font-medium text-fg">React Aria</span> and{' '}
-              <BaseUiIcon className="inline-flex h-[calc(0.85em+1px)] w-auto translate-y-[-0.08em]" />{' '}
-              <span className="font-medium text-fg">Base&nbsp;UI</span>.
-              {/* Install with <ShadcnIcon  className="inline-flex h-[calc(0.85em+1px)] w-auto translate-y-[-0.08em]" /> Shadcn CLI */}
-            </p>
-            <div className="flex w-full flex-col gap-2 pt-1 sm:w-auto sm:flex-row sm:items-center sm:gap-3">
-              <LinkButton href="/create" variant="primary" size="lg">
-                Launch the editor
-              </LinkButton>
-              <LinkButton href="/docs/components" variant="default" size="lg">
-                View components
-              </LinkButton>
-            </div>
-          </div>
-        </section>
-
-        <section className="mt-16">
-          <Cards />
-        </section>
-
-        {/* Built on modern tools — pulled up to sit on the cards' bottom fade. The
-				    masonry's tallest column leaves a tall, masked-to-black tail below the
-				    other columns; without a big pull-up that reads as dead space. We pull
-				    this row up to where the cards are nearly gone (~15% opacity) so it sits
-				    right on the fade instead. The xl grid (4 columns) is shorter and denser
-				    at the bottom, so it needs a slightly smaller pull than the 3-col layout. */}
-        <section className="relative z-10 -mt-[380px] py-12 shadow-xs xl:-mt-[330px]">
-          <div className="container flex flex-col items-center justify-center gap-5 lg:gap-10">
-            <h2 className="font-mono text-sm tracking-wide text-pretty text-fg-muted xs:text-base lg:text-base">
-              Built on modern tools
-            </h2>
-            <div className="flex flex-wrap items-center justify-center gap-4 sm:gap-8">
-              {[
-                {
-                  label: 'Shadcn CLI',
-                  icon: <ShadcnIcon className="size-7 sm:size-9" />,
-                  href: 'https://ui.shadcn.com/docs/cli',
-                },
-                {
-                  label: 'React 19',
-                  icon: <ReactJsIcon className="size-7 sm:size-9" />,
-                  href: 'https://react.dev',
-                },
-                {
-                  label: 'React Aria',
-                  icon: <ReactAriaIcon className="size-7 sm:size-9" />,
-                  href: 'https://react-spectrum.adobe.com/react-aria/index.html',
-                },
-                {
-                  label: 'Base UI',
-                  icon: <BaseUiIcon className="h-7 w-auto sm:h-9" />,
-                  href: 'https://base-ui.com',
-                },
-                {
-                  label: 'TypeScript 5',
-                  icon: <TypeScriptIcon className="size-7 sm:size-9" />,
-                  href: 'https://www.typescriptlang.org/',
-                },
-                {
-                  label: 'Tailwind CSS v4',
-                  icon: <TailwindWordmark className="h-5 sm:h-7" />,
-                  href: 'https://tailwindcss.com',
-                },
-              ].map(({ icon, label, href }) => (
-                <Tooltip key={href}>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    aria-label={label}
-                    className="flex items-center justify-center opacity-60 grayscale-100 transition-opacity hover:opacity-100 hover:grayscale-0"
-                    href={href}
-                  >
-                    {icon}
-                  </a>
-                  <TooltipContent placement="top">{label}</TooltipContent>
-                </Tooltip>
-              ))}
-            </div>
-          </div>
-        </section>
-
-        {/* Call to action */}
-        {/*
-				<section className="container max-w-2xl py-20 sm:py-32">
-					<h2 className="text-pretty font-medium text-2xl tracking-tighter lg:text-3xl xl:text-4xl">
-						Fueled by your <span className="">stars</span>.
-					</h2>
-					<p className="mt-2 text-base text-fg-muted">
-						Our contributors are working hard to make the web a more singular place.
-					</p>
-					<div className="mt-6 flex items-end justify-between gap-6 px-8 sm:items-end sm:justify-between sm:gap-10 lg:gap-20">
-						<div className="flex flex-col items-center gap-6">
-							<svg
-								aria-hidden="true"
-								width="40"
-								viewBox="0 0 236 66"
-								fill="none"
-								xmlns="http://www.w3.org/2000/svg"
-								className="-rotate-110 text-fg-muted"
-							>
-								<path
-									fillRule="evenodd"
-									clipRule="evenodd"
-									d="M23.9244 0.127533C18.3186 0.573079 9.85122 2.47255 4.6676 4.44233C1.59496 5.61482 0 7.46734 0 9.81232C0 11.2897 0.422197 12.298 1.43077 13.2594C2.39244 14.174 3.02573 14.4085 5.74654 14.9713C10.4611 15.9327 20.4765 20.4351 27.5365 24.7968C31.5708 27.2825 32.8374 27.8218 32.8374 27.0479C32.8374 26.6727 31.0313 24.5154 27.513 20.6931C26.3168 19.3799 25.3317 18.2308 25.3317 18.1605C25.3317 17.9963 26.7625 18.0901 34.362 18.7467C54.9088 20.5055 75.5964 24.7733 95.463 31.3861C103.156 33.9422 107.894 35.7244 116.221 39.1949C123.633 42.2669 125.322 42.8766 126.846 42.8766C127.925 42.8766 128.183 42.7828 128.957 42.0793C129.731 41.3523 129.825 41.1413 129.802 40.2033C129.778 39.0777 129.848 39.1949 127.034 35.7244C125.298 33.5904 124.078 31.4096 124.078 30.4247C124.078 29.4398 124.829 29.0412 126.893 29.0412C131.584 29.0412 140.802 32.5587 157.338 40.6723C166.72 45.2685 169.605 46.5816 170.708 46.7692C172.514 47.0506 173.944 46.1361 174.367 44.4477C174.46 44.049 174.226 42.0089 173.804 39.4294C173.053 34.974 173.03 33.8249 173.71 32.5117C174.179 31.6207 174.718 31.3861 176.313 31.3861C180.911 31.3861 190.762 36.0761 205.562 45.2919C214.85 51.0605 219.541 54.3435 227.845 60.886C230.542 63.0199 233.052 64.9194 233.427 65.107C234.318 65.5291 235.35 65.1539 235.726 64.2159C236.218 63.0434 235.515 62.1523 231.175 58.2596C224.537 52.3034 220.104 48.8328 212.974 44.0022C198.854 34.4112 187.22 28.1266 180.488 26.4617C171.599 24.234 166.908 28.9943 168.855 38.1632C169.136 39.4998 169.347 40.6019 169.3 40.6488C169.253 40.6957 167.424 39.7343 165.242 38.5149C159.59 35.3492 148.777 30.0261 144.367 28.2204C137.542 25.383 133.39 24.1167 129.567 23.6477C124.782 23.0615 120.889 24.703 119.716 27.7749C119.223 29.0881 119.364 31.5503 120.044 33.098C120.302 33.7077 120.49 34.2001 120.443 34.2001C120.396 34.2001 118.262 33.1683 115.681 31.902C90.4436 19.4267 60.1159 11.8525 28.9907 10.2579C22.7516 9.9296 22.9158 10.0468 26.6452 8.63983C29.366 7.63148 30.7264 6.0838 30.7264 4.06711C30.703 3.26982 29.882 1.72214 29.1315 1.089C28.4043 0.479303 26.5279 -0.0835156 25.4959 0.0102836C25.2144 0.0337334 24.5108 0.080633 23.9244 0.127533Z"
-									fill="currentColor"
-								/>
-							</svg>
-							<div className="flex items-center gap-1">
-								{contributors.map((contributor) => (
-									<Tooltip key={contributor.login}>
-										<a href={contributor.html_url} target="_blank" rel="noopener noreferrer">
-											<Avatar>
-												<AvatarImage src={contributor.avatar_url} alt={contributor.login} />
-												<AvatarFallback>{contributor.login?.charAt(0).toUpperCase()}</AvatarFallback>
-											</Avatar>
-										</a>
-										<TooltipContent>{contributor.login}</TooltipContent>
-									</Tooltip>
-								))}
-							</div>
-						</div>
-						<LinkButton size="lg" className="h-12" href={siteConfig.links.github} target="_blank">
-							<GitHubIcon />
-							Star on GitHub
-						</LinkButton>
-					</div>
-				</section> */}
-      </main>
-      <Footer />
-    </div>
-  )
-}
